+++ app/Http/Controllers/QuickSaleController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers;
 use App\Models\Product;
 use App\Models\Vat;
 use App\Models\Discount;
+use App\Models\OrderState;
 use App\Models\QuickSale;
 use App\Models\ExerciseProduct;
 use Illuminate\Http\Request;
@@ -39,8 +40,9 @@ class QuickSaleController extends Controller
         $products = Product::where('stock_quantity', '>', 0)->get()->pluck(null, 'id');
         $vats = Vat::all()->pluck(null, 'id');
         $discounts = Discount::all()->pluck(null, 'id');
+        $orderStates = OrderState::all()->pluck(null, 'id');
 
-        return view('quick_sales.create', compact('products', 'vats', 'discounts'));
+        return view('quick_sales.create', compact('products', 'vats', 'discounts', 'orderStates'));
     }
 
     /**
@@ -49,11 +51,9 @@ class QuickSaleController extends Controller
      * @param  \Illuminate\Http\Request  $request
      * @return \Illuminate\Http\Response
      */
-    public function store(Request $request)
+    public function store(StoreQuickSaleRequest $request)
     {
         if ($request->isMethod('POST')) {
-
-            $this->_validateRequest($request);
 
             $product = Product::findOrFail($request->product_id);
 
@@ -67,8 +67,8 @@ class QuickSaleController extends Controller
 
                 DB::beginTransaction();
 
-                $quickSale = quickSale::create(array_merge(
-                    $request->all(),
+                $quickSale = QuickSale::create(array_merge(
+                    $request->validated(),
                     [
                         'selling_price' => $product->selling_price,
                     ]
@@ -82,14 +82,16 @@ class QuickSaleController extends Controller
 
                 DB::commit();
 
-                $this->updateStaffStatusBarInfo(
-                    $quickSale->totalTTC(),
-                    '+'
-                );
+                if ($quickSale->paid) {
+                    $this->updateStaffStatusBarInfo(
+                        $quickSale->totalTTC(),
+                        '+'
+                    );
+                }
 
                 session()->flash('success', "Donnée enregistrée");
 
-                return redirect()->to(route('quick_sale.printing.receipt', ['quick_sale' => $quickSale]));    //to or away
+                return redirect()->to(route('quick_sale.printing.one', ['quick_sale' => $quickSale]));    //to or away
 
             } catch (\Exception $ex) {
                 DB::rollBack();
@@ -127,8 +129,9 @@ class QuickSaleController extends Controller
         $products = Product::where('stock_quantity', '>', 0)->get()->pluck(null, 'id');
         $vats = Vat::all()->pluck(null, 'id');
         $discounts = Discount::all()->pluck(null, 'id');
+        $orderStates = OrderState::all()->pluck(null, 'id');
 
-        return view('quick_sales.edit', compact('quickSale', 'products', 'vats', 'discounts'));
+        return view('quick_sales.edit', compact('quickSale', 'products', 'vats', 'discounts', 'orderStates'));
     }
 
     /**
@@ -138,11 +141,12 @@ class QuickSaleController extends Controller
      * @param  \App\Models\QuickSale  $quickSale
      * @return \Illuminate\Http\Response
      */
-    public function update(Request $request, QuickSale $quickSale)
+    public function update(UpdateQuickSaleRequest $request, QuickSale $quickSale)
     {
         if ($request->isMethod('PUT')) {
 
-            $this->_validateRequest($request);
+            //Gate
+            $this->authorize('cudQuickSale', $quickSale);
 
             $product = Product::findOrFail($quickSale->product_id);
 
@@ -158,12 +162,14 @@ class QuickSaleController extends Controller
                     'stock_quantity' => $product->stock_quantity + $quickSale->quantity,
                 ]);
 
-                $this->updateStaffStatusBarInfo(
-                    $quickSale->totalTTC(),
-                    '-'
-                );
+                if ($quickSale->paid) {
+                    $this->updateStaffStatusBarInfo(
+                        $quickSale->totalTTC(),
+                        '-'
+                    );
+                }
 
-                $quickSale->update($request->all());
+                $quickSale->update($request->validated());
 
                 $product->update([
                     'stock_quantity' => $product->stock_quantity - $quickSale->quantity,
@@ -171,10 +177,12 @@ class QuickSaleController extends Controller
 
                 DB::commit();
 
-                $this->updateStaffStatusBarInfo(
-                    $quickSale->totalTTC(),
-                    '+'
-                );
+                if ($quickSale->paid) {
+                    $this->updateStaffStatusBarInfo(
+                        $quickSale->totalTTC(),
+                        '+'
+                    );
+                }
 
                 session()->flash('success', "Donnée enregistrée");
 
@@ -198,27 +206,12 @@ class QuickSaleController extends Controller
      */
     public function destroy(QuickSale $quickSale)
     {
+        //Gate
+        $this->authorize('cudQuickSale', $quickSale);
+
         $quickSale->delete();
 
         return back()->withDanger('Donnée supprimée');
-    }
-
-    /**
-     * validateRequest
-     *
-     * Validate creation and edition incomming data
-     *
-     * @param mixed $request
-     * @return void
-     */
-    private function _validateRequest($request)
-    {
-        $request->validate([
-            'product_id' => ['required'],
-            'vat_id' => ['nullable'],
-            'discount_id' => ['nullable'],
-            'quantity' => ['required'],
-        ]);
     }
 
     public function printingAll(Request $request)
@@ -240,14 +233,5 @@ class QuickSaleController extends Controller
         $pdf->save(public_path("libraries/docs/quick_sale_{$quickSale->id}.pdf"));
 
         return $pdf->stream("quick_sale_{$quickSale->id}.pdf");
-    }
-
-    public function receipt(Request $request, QuickSale $quickSale)
-    {
-        $pdf = PDF::loadView('quick_sales.printing.receipt', compact('quickSale'));
-        $pdf->setPaper('a4', 'portrait');
-        $pdf->save(public_path("libraries/docs/receipt_{$quickSale->id}.pdf"));
-
-        return $pdf->stream("receipt_{$quickSale->id}.pdf");
     }
 }
